Suggest a change in project structure
