--- src/main/java/com/PrincessCyphers/WebApp/MainController.java
@@ -1,13 +1,0 @@
-package com.PrincessCyphers.WebApp;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-    @GetMapping(" ")
-    public String showHomePage() {
-        return "index";
-    }
-}
