--- src/main/java/com/PrincessCyphers/user/UserRepository.java
@@ -1,8 +1,0 @@
-// interface for the user
-
-package com.PrincessCyphers.user;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Integer> {
-}
