+++ src/main/java/com/learning/WebAppApplication.java
@@ -1,4 +1,4 @@
-package com.PrincessCyphers.WebApp;
+package com.learning;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
