+++ src/main/java/com/learning/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.learning.repository;
+
+import com.learning.user.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Integer> { }
