+++ src/main/resources/application.properties
@@ -1,7 +1,12 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/usersdb
-spring.datasource.username=root
-spring.datasource.password=password
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL55Dialect
-spring.jpa.hibernate.hbm2ddl.auto=update
-spring.jpa.show_sql=true
-
+#spring.datasource.url=jdbc:mysql://localhost:3306/usersdb
+#spring.datasource.username=root
+#spring.datasource.password=password
+#spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL55Dialect
+#spring.jpa.hibernate.hbm2ddl.auto=update
+#spring.jpa.show_sql=true
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=accounts
+spring.datasource.password=accounts
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=update
