+++ src/test/java/com/learning/repository/UserRepositoryTests.java
@@ -1,7 +1,6 @@
-package com.PrincessCyphers.WebApp;
+package com.learning.repository;
 
-import com.PrincessCyphers.user.User;
-import com.PrincessCyphers.user.UserRepository;
+import com.learning.user.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,22 +9,22 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.annotation.Rollback;
 
 @DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)            //test will be against real db
-@Rollback(false)                 // Keeps data committed to db
+//@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)            //test will be against real db
+//@Rollback(false)                 // Keeps data committed to db
+public class UserRepositoryTests {
 
-
-public class UserRepoTests {
-    //ref to user repo interface
-    @Autowired private UserRepository repo;
+    @Autowired
+    private UserRepository repository;
 
     @Test
     public void testAddNew() {
-        User user = new User();
+        final User user = new User();
         user.setEmail("newuser@example.com");
         user.setPassword("1234");
         user.setFirstName("New");
         user.setLastName("User");
-        User savedUser = repo.save(user);
+
+        final User savedUser = repository.save(user);
 
         Assertions.assertThat(savedUser).isNotNull();
         Assertions.assertThat(savedUser.getId()).isGreaterThan(0);
